requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
billiard==4.1.0
celery==5.3.0
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.2.2
django-celery-results==2.5.1
django-jazzmin==2.6.0
django-tinymce==3.6.1
exceptiongroup==1.1.1
h11==0.14.0
idna==3.4
instaloader==4.9.6
kombu==5.3.1
mysqlclient==2.2.0
outcome==1.2.0
prompt-toolkit==3.0.38
PySocks==1.7.1
python-dateutil==2.8.2
redis==4.5.5
requests==2.31.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.4
trio==0.22.0
trio-websocket==0.10.3
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
vine==5.0.0
wcwidth==0.2.6
wsproto==1.2.0
psycopg2-binary==2.9.1
gunicorn==20.1.0
whitenoise==6.5.0
 